feat: cria validators para alguns endpoints
